requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
attrs==24.3.0
    # via
    #   cattrs
    #   jsii
aws-cdk-lib==2.175.1
    # via
    #   -r requirements.in
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-authorizers-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
aws-cdk.asset-awscli-v1==2.2.218
    # via aws-cdk-lib
aws-cdk.asset-kubectl-v20==2.1.3
    # via aws-cdk-lib
aws-cdk.asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk.aws-apigatewayv2-alpha==2.114.1.a0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-apigatewayv2-authorizers-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-apigatewayv2-authorizers-alpha==2.114.1a0
    # via -r requirements.in
aws-cdk.aws-apigatewayv2-integrations-alpha==2.114.1a0
    # via -r requirements.in
aws-cdk.aws-lambda-python-alpha==2.175.1a0
    # via -r requirements.in
aws-cdk.cloud-assembly-schema==39.1.44
    # via aws-cdk-lib
cattrs==24.1.2
    # via jsii
constructs==10.4.2
    # via
    #   -r requirements.in
    #   aws-cdk-lib
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-authorizers-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
exceptiongroup==1.2.2
    # via cattrs
importlib-resources==6.5.2
    # via jsii
jsii==1.106.0
    # via
    #   aws-cdk-lib
    #   aws-cdk.asset-awscli-v1
    #   aws-cdk.asset-kubectl-v20
    #   aws-cdk.asset-node-proxy-agent-v6
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-authorizers-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   aws-cdk.cloud-assembly-schema
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.asset-awscli-v1
    #   aws-cdk.asset-kubectl-v20
    #   aws-cdk.asset-node-proxy-agent-v6
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-authorizers-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   aws-cdk.cloud-assembly-schema
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via jsii
six==1.17.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.asset-awscli-v1
    #   aws-cdk.asset-kubectl-v20
    #   aws-cdk.asset-node-proxy-agent-v6
    #   aws-cdk.aws-apigatewayv2-alpha
    #   aws-cdk.aws-apigatewayv2-authorizers-alpha
    #   aws-cdk.aws-apigatewayv2-integrations-alpha
    #   aws-cdk.aws-lambda-python-alpha
    #   aws-cdk.cloud-assembly-schema
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via
    #   cattrs
    #   jsii
zipp==3.21.0
    # via importlib-resources
